Useing 2 pools instead of request string
Made some changes to the pools, now pools can be setup to
use or not use the gpu and you can choose the starting index of the
processes id the pool creates.
Changed Signify! Daniel please check. small change and everything
seems to work but still

This commit relies on the commit Or wrote for the camera changes!